refactor(chat): ChatController를 역할별로 분리 (Room, Message, Member)
기존의 비대한 ChatController를 도메인 역할에 따라 3개의 컨트롤러로 분리하여 유지보수성을 향상시킴.

- ChatController 제거 및 아래 컨트롤러로 기능 이관
  1. ChatRoomController: 채팅방 생성, 조회, 입장/퇴장 관리
  2. ChatMessageController: 메시지 조회, 검색, 파일 업로드 URL 발급
  3. ChatMemberController: 참여자 조회, 차단, 신고, 알림/번역 설정
- ChatWebSocketController: 실시간 소켓 통신 역할 유지
- Swagger(@Tag, @Operation) 설정을 각 컨트롤러 성격에 맞게 재구성
- 기존 API 엔드포인트(/api/v1/chat)는 유지하여 하위 호환성 보장
